Watch remove channel, update state accordingly
In order to remember what was removed, we now store changes
to the data store (since the last time persist was called).

At the moment, we only expose the most recent change
to the persister you define in your app code. So if that persister
needs the id of the thing that was deleted, don't call remove more
than once in between persist calls.

In the future, we want to iterate through all the
changes since the last persist.